Switch inih to It's Official Repository @ benhoyt/inih
## Info
This Commit switches the Inih Submodule from it's mirror repository svn2github/inih to it's official repo at benhoyt/inih

## Reasoning
As explained in Info, svn2github is a Mirror Host, which was used mainly as there was no official repo for inih at the Time it was originally added. However, now that there is a Official repo, svn2github is considered unnecessary as well as inconsistent. This commit will make it easier to manage this submodule in the future, as well as able to fix issues related to inih, should any appear in the foreseeable future.