citra-room: Fix always false case in main() related to port range
If the variable we're checking is a u16, then there can never be values
outside of the 0-65535 range. This is bad because an arbitrary larger
value can be truncated down into a valid value, making an otherwise
malformed argument well-formed.

Change it to use u32 to allow the check to function properly.